+++ packages/host-papp-react-ui/.storybook/preview.tsx
@@ -1,4 +1,4 @@
-import { ThemeProvider, Toaster, defaultTheme } from '@novasamatech/tr-ui';
+import { ThemeProvider, Toaster, berlinTheme } from '@novasamatech/tr-ui';
 import type { Preview } from '@storybook/react';
 
 // @ts-expect-error css imports are not defined here
@@ -7,7 +7,7 @@ import '@novasamatech/tr-ui/styles.css';
 const preview: Preview = {
   decorators: [
     Story => (
-      <ThemeProvider defaultMode="light" theme={defaultTheme}>
+      <ThemeProvider defaultMode="light" theme={berlinTheme}>
         <Story />
         <Toaster />
       </ThemeProvider>